Add utility functions bits_to_target and target_to_bits to cryptocur.py
