+++ lib/chains/mazacoin.py
@@ -1,5 +1,5 @@
 '''Chain-specific Mazacoin code'''
-from cryptocur import CryptoCur, hash_encode, hash_decode, rev_hex, int_to_hex
+from cryptocur import CryptoCur, hash_encode, hash_decode, rev_hex, int_to_hex, bits_to_target, target_to_bits
 import os
 
 from coinhash import SHA256dHash
@@ -105,32 +105,6 @@ class Mazacoin(CryptoCur):
         f.seek(height*80)
         h = f.write(data)
         f.close()
-
-    def bits_to_target(self, bits):
-        MM = 256*256*256
-        a = bits%MM
-        if a < 0x8000:
-            a *= 256
-        target = (a) * pow(2, 8 * (bits/MM - 3))
-        return target
-
-    def target_to_bits(self, target):
-        MM = 256*256*256
-        c = ("%064X"%target)[2:]
-        i = 31
-        while c[0:2]=="00":
-            c = c[2:]
-            i -= 1
-
-        c = int('0x'+c[0:6],16)
-        if c >= 0x800000:
-            c /= 256
-            i += 1
-
-        new_bits = c + MM * i
-        return new_bits
-
-
 
     def get_target_v1(self, block_height, chain=None):
         # params
@@ -239,9 +213,9 @@ class Mazacoin(CryptoCur):
 
             if CountBlocks <= PastBlocksMin:
                 if CountBlocks == 1:
-                    PastDifficultyAverage = self.bits_to_target(BlockReading.get('bits'))
+                    PastDifficultyAverage = bits_to_target(BlockReading.get('bits'))
                 else:
-                    bnNum = self.bits_to_target(BlockReading.get('bits'))
+                    bnNum = bits_to_target(BlockReading.get('bits'))
                     PastDifficultyAverage = ((PastDifficultyAveragePrev * CountBlocks)+(bnNum)) / (CountBlocks + 1)
                 PastDifficultyAveragePrev = PastDifficultyAverage
 
@@ -268,7 +242,7 @@ class Mazacoin(CryptoCur):
 
         bnNew = min(bnNew, max_target)
 
-        new_bits = self.target_to_bits(bnNew)
+        new_bits = target_to_bits(bnNew)
         return new_bits, bnNew
 
     def get_target(self, block_height, chain=None):
